+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/items/FlowerHeadwearHelmet.java
@@ -1,7 +1,10 @@
 package com.telepathicgrunt.the_bumblezone.items;
 
 import com.telepathicgrunt.the_bumblezone.mixin.effects.MobEffectInstanceAccessor;
+import com.telepathicgrunt.the_bumblezone.modcompat.ModChecker;
+import com.telepathicgrunt.the_bumblezone.modcompat.ModCompat;
 import com.telepathicgrunt.the_bumblezone.modinit.BzEffects;
+import com.telepathicgrunt.the_bumblezone.modinit.BzItems;
 import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
@@ -12,7 +15,9 @@ import net.minecraft.world.item.DyeableLeatherItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
-public class FlowerHeadwearHelmet extends BzDyeableArmor implements DyeableLeatherItem {
+import java.util.concurrent.atomic.AtomicReference;
+
+public class FlowerHeadwearHelmet extends BzDyeableArmor implements DyeableLeatherItem, FlowerHeadwearHelmetMixinTarget {
     public FlowerHeadwearHelmet(ArmorMaterial material, Type armorType, Properties properties) {
         super(material, armorType, properties);
     }
@@ -47,12 +52,23 @@ public class FlowerHeadwearHelmet extends BzDyeableArmor implements DyeableLeath
         }
     }
 
-    public static ItemStack getFlowerHeadware(Entity entity) {
+    public static ItemStack getFlowerHeadwear(Entity entity) {
         for(ItemStack armor : entity.getArmorSlots()) {
             if(armor.getItem() instanceof FlowerHeadwearHelmet) {
                 return armor;
             }
         }
-        return ItemStack.EMPTY;
+
+        AtomicReference<ItemStack> flowerStack = new AtomicReference<>(ItemStack.EMPTY);
+        for (ModCompat compat : ModChecker.CUSTOM_EQUIPMENT_SLOTS_COMPATS) {
+            compat.getNumberOfMatchingEquippedItemsInCustomSlots(entity, (itemStack) -> {
+                if (itemStack.is(BzItems.FLOWER_HEADWEAR.get())) {
+                    flowerStack.set(itemStack);
+                    return true;
+                }
+                return false;
+            });
+        }
+        return flowerStack.get();
     }
 }