+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/items/FlowerHeadwearHelmetMixinTarget.java
@@ -1,0 +1,3 @@
+package com.telepathicgrunt.the_bumblezone.items;
+
+public interface FlowerHeadwearHelmetMixinTarget { }