+++ fabric-base/src/main/java/com/telepathicgrunt/the_bumblezone/fabricbase/FabricClientBaseEventManager.java
@@ -38,10 +38,8 @@ import net.minecraft.client.gui.screens.inventory.MenuAccess;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
 import net.minecraft.client.renderer.item.ItemProperties;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.MenuType;
-import org.intellij.lang.annotations.Identifier;
 
 public class FabricClientBaseEventManager {